move error message location
